Правки инспекции кода
