+++ 2020-11/spring-04/.gitignore
@@ -8,6 +8,7 @@ target/
 .settings
 .springBeans
 .sts4-cache
+.log
 
 ### IntelliJ IDEA ###
 .idea
