+++ 2020-11/spring-04/spring-04-exercise/spring-shell.log
@@ -1,119 +1,5 @@
-1608394878591:hello-to
-1608394896388:qqq
-1608395710062:s
-1608395723706:surname
-1608395742915:surname Tsaplin
-1608395749681:n Sergey
-1608395758318:n SSS
-1608395775343:quit
-1608396086723:help
-1608396309739:s
-1608396316980:s Tsaplin
-1608396322380:n Sergey
-1608396329138:y
-1608396337146:exit
-1608396530301:s Tsaplin
-1608396536073:n Srg
-1608396542790:exit
-1608396939375:s Sts
-1608396954973:n S
-1608396973786:exit
-1608397215043:s Tsaplin
-1608397227918:n Sergey
-1608397240157:exit
-1608397268143:help
-1608397295467:s Tsaplin
-1608397300123:n Sergey
-1608397314770:history
-1608397322777:script
-1608397347639:exit
-1608397599894:s Tsaplin
-1608397607195:n Sergey
-1608397618570:exit
-1608397707514:s Tsaplin
-1608397712214:n Sergey
-1608397716181:exit
-1608397788212:s Tsaplin
-1608397796786:n Sergey
-1608397802969:exit
-1608397929110:s Tsaplin
-1608397932483:n Sergey
-1608397936626:exit
-1608401025906:s Tsaplin
-1608401030193:n Sergey
-1608401039272:continue
-1608401045383:quit
-1608478561897:n
-1608478566158:n Sergey
-1608478571426:s Tsaplin
-1608478581233:a yyy
-1608478615486:exit
-1608478682336:s Tsaplin
-1608478686694:n Sergey
-1608478695939:help
-1608478711919:a y
-1608478733585:a yy
-1608478756471:a yyy
-1608478777701:a yyyy
-1608478786754:a yyyyy
-1608478797388:a ttt
-1608478804647:exit
-1608479221276:help
-1608479258026:exit
-1608479575530:help
-1608479624342:s Tsaplin
-1608479629478:n Sergey
-1608479634853:exec
-1608479735917:s Tsaplin
-1608479739013:n Sergey
-1608479741762:exec
-1608479875539:help
-1608479890179:exec
-1608479906722:exit
-1608480047899:s Ivanov
-1608480058241:exit
-1608480121278:s sss
-1608480126941:exit
-1608480513571:help
-1608480519242:exit
-1608480568722:s Ivanov
-1608480573184:n Petr
-1608480577415:exec
-1608480596382:exit
-1608480824125:s Ivanov
-1608480828867:n Sergey
-1608480830906:exec
-1608480843150:exit
-1608480886977:s Ivanocv
-1608480891864:n Petr
-1608480894619:exec
-1608480909774:exit
-1608481916639:s Petrov
-1608481921588:n Ivan
-1608481924556:exec
-1608481933579:exit
-1608482051270:exec
-1608482068876:exit
-1608482100652:exec
-1608482122498:exit
-1608482178997:exec
-1608482187216:n Sergey
-1608482190278:exec
-1608482202495:s tsaplin
-1608482205437:exec
-1608482223737:exit
-1608483948648:help
-1608483984741:n vera
-1608483999772:e
-1608484023797:s tsaplina
-1608484028011:e
-1608484090952:exit
-1608485213802:s Tsaplin
-1608485216944:n sergey
-1608485220575:e
-1608485239582:exit
-1608485455437:s Tsaplin
-1608485457331:exec
-1608485466363:exit
-1608485671932:s Tsaplin
-1608485674881:exec
+1608668901449:s Tsaplin
+1608668904281:exec
+1608668912492:exit
+1608668997533:exec
+1608669003193:exit
