--- 2020-11/spring-04/spring-04-exercise/src/main/java/ru/otus/spring/service/IMyCSVFile.java
@@ -1,10 +1,0 @@
-package ru.otus.spring.service;
-
-import java.io.IOException;
-import java.util.List;
-
-public interface IMyCSVFile {
-    void setFileName(String fileName);
-    List<IQuestionLine> readCSVFile() throws IOException;
-
-}
