+++ 2020-11/spring-04/spring-04-exercise/src/main/java/ru/otus/spring/service/QuestionnaireImpl.java
@@ -16,7 +16,7 @@ import java.util.Scanner;
 @RequiredArgsConstructor
 public class QuestionnaireImpl implements Questionnaire {
     private final User user;
-    private final IMyCSVFile myFile;
+    private final MyCSVFile myFile;
     private IQuestionLine qLine;
 
     @Override
